Second update deprecating SArray which is no longer used by functions and distributions.
The need to have multiple chains requires a separation between the value of
a node and its dimension.  SArray objects, which combined both of these,
could not be readily adapted to multiple chains without introducing a lot
of redundancy.

Under the new API, functions and distributions have direct access to
the values of parameters (which is also more efficient).   Dimension
information is provided in a separate argument, when necessary.

The DistReal and DistDiscrete classes are now deprecated. A new class
DistScalarRmath is introduced for wrappers around the Rmath d-p-q-r
functions. The DistScalar class is reserved for scalar distributions with
scalar parameters.

The command line interface now allows you to run JAGS in a debugger
(gdb or valgrind).