Modified Node class so that the parents are constant pointers and children are contained in an external dynamically allocated set.
Propagating this change was mostly straightforward, except for
MixtureNode creation, where it entailed fairly major changes to the
Compiler.  A compiler now builds the model graph on-the-fly. The graph
is owned by the model, and other objects used by the compiler which
previously had their own graphs, now merely contain a reference to the
model graph. Any function which might dynamically allocate a new node,
now needs an explicit reference to the graph that will contain it.